refactor: 컴포넌트별 폴더 생성
