+++ index.html
@@ -7,14 +7,15 @@
   <title>Document</title>
   <link href="./reset.style.css" rel="stylesheet">
   <link href="./style.css" rel="stylesheet">
-  <link href="./LottoGameCard.style.css" rel="stylesheet">
-  <link href="./PurchaseAmountForm.style.css" rel="stylesheet">
-  <link href="./LottoInfo.style.css" rel="stylesheet">
-  <link href="./WinningNumbersAndBonusNumberForm.style.css" rel="stylesheet">
-  <link href="./Modal.style.css" rel="stylesheet">
-  <link href="./LottoRanksTable.style.css" rel="stylesheet">
-  <link href="./LottoResult.style.css" rel="stylesheet">
-  <link href="./LottoResultModal.style.css" rel="stylesheet">
+  <link href="src/step-2/LottoGameCard/LottoGameCard.style.css" rel="stylesheet">
+  <link href="src/step-2/PurchaseAmountForm/PurchaseAmountForm.style.css" rel="stylesheet">
+  <link href="src/step-2/LottoInfo/LottoInfo.style.css" rel="stylesheet">
+  <link href="src/step-2/WinningNumbersAndBonusNumber/WinningNumbersAndBonusNumberForm.style.css"
+        rel="stylesheet">
+  <link href="src/step-2/Modal/Modal.style.css" rel="stylesheet">
+  <link href="src/step-2/LottoRanksTable/LottoRanksTable.style.css" rel="stylesheet">
+  <link href="src/step-2/LottoResult/LottoResult.style.css" rel="stylesheet">
+  <link href="src/step-2/LottoResultModal/LottoResultModal.style.css" rel="stylesheet">
 </head>
 
 <body>
