+++ src/step-2/LottoGameCard/LottoGameCard.style.css
