+++ src/step-2/LottoInfo/LottoInfo.style.css
