+++ src/step-2/LottoRanksTable/LottoRanksTable.style.css
