+++ src/step-2/LottoResult/LottoResult.style.css
