+++ src/step-2/LottoResultModal/LottoResultModal.style.css
