+++ src/step-2/PurchaseAmountForm/PurchaseAmountForm.style.css
